#267: Moved the files for DH and TDH bonuses to the model folder
These are still Parsers for now.
