+++ src/app/ffbe/model/effects/passives/stats/passive-equipment-stats-doublehand-increase.parser.ts
@@ -1,6 +1,6 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
-import {HTML_LINE_RETURN} from '../skill-effects.mapper';
+import {EffectParser} from '../../../../mappers/effects/effect-parser';
+import {Skill} from '../../../skill.model';
+import {HTML_LINE_RETURN} from '../../../../mappers/effects/skill-effects.mapper';
 
 export class PassiveEquipmentStatsDoublehandIncreaseParser extends EffectParser {
   public parse(effect: Array<any>, skill: Skill): string {
